backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
psycopg2-binary==2.9.9
requests==2.31.0
beautifulsoup4==4.12.0
httpx==0.25.2
playwright==1.40.0
exa-py==1.0.32
